move out GlyphData and GlyphInfo
