+++ packages/flame/lib/src/text/common/glyph_info.dart
@@ -1,0 +1,15 @@
+
+/// Helper class that stores dimensions of a single glyph for a spritesheet-
+/// based font.
+class GlyphInfo {
+  double srcLeft = 0;
+  double srcTop = 0;
+  double srcRight = 0;
+  double srcBottom = 0;
+  double rstSCos = 1;
+  double rstSSin = 0;
+  double rstTx = 0;
+  double rstTy = 0;
+  double width = 0;
+  double height = 0;
+}
