+++ packages/flame/lib/src/text/formatters/sprite_font_text_formatter.dart
@@ -1,6 +1,8 @@
 import 'dart:typed_data';
 import 'dart:ui' hide LineMetrics;
 
+import 'package:flame/src/text/common/glyph_data.dart';
+import 'package:flame/src/text/common/glyph_info.dart';
 import 'package:flame/src/text/common/line_metrics.dart';
 import 'package:flame/src/text/formatters/text_formatter.dart';
 import 'package:flame/src/text/inline/sprite_font_text_element.dart';
