+++ packages/flame/lib/text.dart
@@ -1,3 +1,4 @@
+export 'src/text/common/glyph_data.dart' show GlyphData;
 export 'src/text/sprite_font_renderer.dart' show SpriteFontRenderer, GlyphData;
 export 'src/text/text_paint.dart' show TextPaint;
 export 'src/text/text_renderer.dart' show TextRenderer;
